@@ -16,14 +16,18 @@
     <properties>
         <maven-checkstyle-plugin.version>2.17</maven-checkstyle-plugin.version>
         <atlasmap.version>2.1.0-SNAPSHOT</atlasmap.version>
+        <assertj.version>3.11.1</assertj.version>
         <junit.version>4.12</junit.version>
-        <spring-boot.version>1.5.12.RELEASE</spring-boot.version>
-        <springframework.version>4.3.18.RELEASE</springframework.version>
-        <selenide.version>4.12.2</selenide.version>
+        <spring-boot.version>2.3.2.RELEASE</spring-boot.version>
+        <selenide.version>5.12.2</selenide.version>
         <selenium.version>3.13.0</selenium.version>
         <node.version>v8.11.4</node.version>
         <yarn.version>v1.9.4</yarn.version>
-        <cucumber.version>4.7.1</cucumber.version>
+        <guava.version>25.1-jre</guava.version>
+        <cucumber.version>6.2.2</cucumber.version>
+        <failsafe.plugin.version>2.22.0</failsafe.plugin.version>
+        <cluecumber.plugin.version>2.5.0</cluecumber.plugin.version>
+        <spring.version>5.2.5.RELEASE</spring.version>
         <log4j.version>2.9.1</log4j.version>
         <httpcomponents.version>4.5.3</httpcomponents.version>
         <maven.compiler.source>1.8</maven.compiler.source>
