@@ -1,8 +1,11 @@
 package io.atlasmap.qe.test.atlas.utils;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -13,27 +16,20 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import cucumber.api.Result;
-import cucumber.api.Result.Type;
-import cucumber.api.TestCase;
-import cucumber.api.event.EventListener;
-import cucumber.api.event.EventPublisher;
-import cucumber.api.event.TestCaseFinished;
-import cucumber.api.event.TestRunFinished;
-import cucumber.api.event.TestSourceRead;
-import gherkin.AstBuilder;
-import gherkin.Parser;
-import gherkin.ParserException;
-import gherkin.ast.Comment;
-import gherkin.ast.Feature;
-import gherkin.ast.GherkinDocument;
-
+import io.cucumber.core.gherkin.messages.internal.gherkin.GherkinDocumentBuilder;
+import io.cucumber.core.gherkin.messages.internal.gherkin.Parser;
+import io.cucumber.messages.IdGenerator;
+import io.cucumber.messages.Messages;
+import io.cucumber.plugin.EventListener;
+import io.cucumber.plugin.event.EventPublisher;
+import io.cucumber.plugin.event.Status;
+import io.cucumber.plugin.event.TestCase;
+import io.cucumber.plugin.event.TestCaseFinished;
+import io.cucumber.plugin.event.TestRunFinished;
+import io.cucumber.plugin.event.TestSourceRead;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 
 @Slf4j
 public class MailFormatter implements EventListener {
@@ -44,8 +40,8 @@ public class MailFormatter implements EventListener {
 
     private ObjectMapper mapper = new ObjectMapper();
 
-    private Map<String, String> sustainers = new HashMap<>();
-    private Map<String, Feature> features = new HashMap<>();
+    private Map<URI, String> sustainers = new HashMap<>();
+    private Map<URI, Messages.GherkinDocument.Feature> features = new HashMap<>();
     private Set<String> recipients = new HashSet<>();
     private List<ScenarioResult> results = new ArrayList<>();
     private int passed, failed, total = 0;
@@ -59,26 +55,25 @@ public class MailFormatter implements EventListener {
         publisher.registerHandlerFor(TestSourceRead.class, this::onTestSourceRead);
         publisher.registerHandlerFor(TestCaseFinished.class, this::onTestCaseFinished);
         publisher.registerHandlerFor(TestRunFinished.class, this::onTestRunFinished);
-        //    publisher.registerHandlerFor(EmbedEvent.class, this::onEmbed);
     }
 
     private void onTestSourceRead(TestSourceRead t) {
-        GherkinDocument doc = parseGherkinSource(t.source);
+        Messages.GherkinDocument doc = parseGherkinSource(t.getSource());
         if (doc != null) {
-            features.put(t.uri, doc.getFeature());
-            for (Comment c : doc.getComments()) {
+            features.put(t.getUri(), doc.getFeature());
+            for (Messages.GherkinDocument.Comment c : doc.getCommentsList()) {
                 Matcher matcher = SUSTAINER_PATTERN.matcher(c.getText());
                 if (matcher.matches()) {
-                    sustainers.put(t.uri, matcher.group(1));
+                    sustainers.put(t.getUri(), matcher.group(1));
                 }
             }
         }
     }
 
     private void onTestCaseFinished(TestCaseFinished t) {
-        String uri = t.getTestCase().getUri();
+        URI uri = t.getTestCase().getUri();
         total++;
-        switch (t.result.getStatus()) {
+        switch (t.getResult().getStatus()) {
             case PASSED:
                 passed++;
                 break;
@@ -90,10 +85,10 @@ public class MailFormatter implements EventListener {
         }
         ;
         results.add(
-                new ScenarioResult(
-                        features.get(uri), t.getTestCase(), t.result.getStatus())
+            new ScenarioResult(
+                features.get(uri), t.getTestCase(), t.getResult().getStatus())
         );
-        if (!t.result.getStatus().equals(Type.PASSED) && sustainers.get(uri) != null) {
+        if (!t.getResult().getStatus().equals(Status.PASSED) && sustainers.get(uri) != null) {
             recipients.add(sustainers.get(uri));
         }
     }
@@ -102,13 +97,13 @@ public class MailFormatter implements EventListener {
         new File(path).mkdirs();
         try (FileWriter out = new FileWriter(new File(path, "report.html"))) {
             out.write("<h2>Atlasmap QE E2E test results</h2>" +
-                    String.format("PASSED %d / %d<br/>\n", passed, total) +
-                    String.format("Failed %d / %d<br/>\n", failed, total));
+                String.format("PASSED %d / %d<br/>\n", passed, total) +
+                String.format("Failed %d / %d<br/>\n", failed, total));
 
             out.write(String.join(
-                    "<br/>\n",
-                    results.stream().map(ScenarioResult::toString).collect(Collectors.toList())
-                    )
+                "<br/>\n",
+                results.stream().map(ScenarioResult::toString).collect(Collectors.toList())
+                )
             );
         } catch (IOException e) {
             log.error("Error writing mail report file", e);
@@ -120,49 +115,42 @@ public class MailFormatter implements EventListener {
         }
     }
 
-    private GherkinDocument parseGherkinSource(String source) {
-        Parser<GherkinDocument> parser = new Parser<>(new AstBuilder());
-        try {
-            return parser.parse(source);
-        } catch (ParserException e) {
-            log.error("Error parsing gherkin source", e);
-        }
-        return null;
+    private Messages.GherkinDocument parseGherkinSource(String source) {
+        return new Parser<>(new GherkinDocumentBuilder(new IdGenerator.Incrementing())).parse(source).build();
     }
 
     @Data
     @RequiredArgsConstructor
     private static class ScenarioResult {
-        private final Feature feature;
+        private final Messages.GherkinDocument.Feature feature;
         private final TestCase testCase;
-        private final Result.Type result;
-        // private final String sustainer;
+        private final Status result;
 
         public String toString() {
             // TODO: consider using a templating engine for this
             StringBuilder sb = new StringBuilder();
 
             sb
-                    .append(feature.getName())
-                    .append(" | ")
-                    .append(testCase.getName())
-                    .append(" | ");
+                .append(feature.getName())
+                .append(" | ")
+                .append(testCase.getName())
+                .append(" | ");
 
-            if (result.equals(Result.Type.PASSED)) {
+            if (result.equals(Status.PASSED)) {
                 sb
-                        .append("<font color=\"green\">")
-                        .append(result)
-                        .append("</font>");
-            } else if (result.equals(Result.Type.SKIPPED)) {
+                    .append("<font color=\"green\">")
+                    .append(result)
+                    .append("</font>");
+            } else if (result.equals(Status.SKIPPED)) {
                 sb
-                        .append("<font color=\"yellow\">")
-                        .append(result)
-                        .append("</font>");
+                    .append("<font color=\"yellow\">")
+                    .append(result)
+                    .append("</font>");
             } else {
                 sb
-                        .append("<font color=\"red\"><b>")
-                        .append(result)
-                        .append("</b></font>");
+                    .append("<font color=\"red\"><b>")
+                    .append(result)
+                    .append("</b></font>");
             }
 
             return sb.toString();
