@@ -2,11 +2,11 @@ package io.atlasmap.qe.test.atlas.steps;
 
 import org.junit.Assert;
 
-import cucumber.api.java.en.And;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
 import io.atlasmap.qe.test.atlas.AtlasmapPage;
 import io.atlasmap.qe.test.atlas.utils.Utils;
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
 
 public class TransformationSteps extends CucumberGlue {
     private AtlasmapPage atlasmapPage = new AtlasmapPage();
